+++ app/static/lib/main.js
@@ -1,6 +1,6 @@
 // mei-friend version and date
 export const version = '0.8.3';
-export const versionDate = '15 Feb 2023';
+export const versionDate = '16 Feb 2023';
 
 var vrvWorker;
 var spdWorker;
@@ -1458,6 +1458,7 @@ export let cmd = {
   delete: () => e.deleteElement(v, cm),
   cmdDelete: () => e.deleteElement(v, cm, true),
   invertPlacement: () => e.invertPlacement(v, cm),
+  betweenPlacement: () => e.invertPlacement(v, cm, true),
   addVerticalGroup: () => e.addVerticalGroup(v, cm),
   toggleStacc: () => e.toggleArtic(v, cm, 'stacc'),
   toggleAccent: () => e.toggleArtic(v, cm, 'acc'),